Move IHDR encoding into own file
